+++ Assets/Scripts/Adapters/Unity/NyrvexaV02MapChrome.cs
@@ -4,7 +4,7 @@
  * Developer:      Irfan Gedik
  * Created Date:   2026-04-26
  * Last Update:    2026-04-26
- * Version:        0.1.8
+ * Version:        0.1.9
  * 
  * Description:
  *   v0.2: sağ üstte çerçeveli minimap (F0 FOW/owner), sağ altta kamera: